refactor(game): move calculations to the collider
