+++ src/game/engine/internals/CollisionDetector.js
@@ -1,5 +1,3 @@
-import { Collision } from '../../shared';
-
 class CollisionDetector {
   /**
    * @type {import('./Collider').Collider[]}
@@ -33,9 +31,12 @@ class CollisionDetector {
   detectCollisions() {
     this.#colliders.forEach((collider1) => {
       this.#colliders.forEach((collider2) => {
-        if (collider1 !== collider2 && collider1.hasCollisionWith(collider2)) {
-          const collision = new Collision(collider1.gameObject, collider2.gameObject);
-          collision.gameObject1.onCollision(collision, collision.gameObject2);
+        if (collider1 !== collider2) {
+          const collision = collider1.hasCollisionWith(collider2);
+
+          if (collision) {
+            collision.gameObject1.onCollision(collision, collision.gameObject2);
+          }
         }
       });
     });
